Support returning null in byzantine fault injection
